--- webtau-core/src/main/resources/webtau.version
@@ -1,1 +1,0 @@
-${project.version}
